+++ spec/logger_spec.rb
@@ -4,10 +4,10 @@ require 'tempfile'
 
 describe ActFluentLoggerRails::Logger do
   before do
-    Rails = double("Rails") unless self.class.const_defined?(:Rails)
-    Rails.stub(env: "test")
-    Rails.stub_chain(:application, :config, :log_level).and_return(:debug)
-    Rails.stub_chain(:application, :config, :log_tags=)
+    stub_const('Rails', Class.new) unless defined?(Rails)
+    allow(Rails).to receive(:env).and_return('test')
+    allow(Rails).to receive_message_chain(:application, :config, :log_level).and_return(:debug)
+    allow(Rails).to receive_message_chain(:application, :config, :log_tags=)
 
     class MyLogger
       attr_accessor :log
@@ -22,7 +22,7 @@ describe ActFluentLoggerRails::Logger do
       end
     end
     @my_logger = MyLogger.new
-    Fluent::Logger::FluentLogger.stub(:new) { @my_logger }
+    allow(Fluent::Logger::FluentLogger).to receive(:new).and_return(@my_logger)
 
     @config_file = Tempfile.new('fluent-logger-config')
     @config_file.close(false)
@@ -36,36 +36,50 @@ EOF
     }
   end
 
+  let(:log_tags) {
+    { uuid: :uuid,
+      foo: ->(request) { 'foo_value' }
+    }
+  }
+
   let(:logger) {
     ActFluentLoggerRails::Logger.new(config_file: File.new(@config_file.path),
-                                     log_tags: {
-                                       uuid: :uuid,
-                                       foo: ->(request) { request.foo }
-                                     })
+                                     log_tags: log_tags)
   }
 
   let(:request) {
-    double('request', uuid: 'uuid_value', foo: 'foo_value')
+    double('request', uuid: 'uuid_value')
   }
 
   describe 'logging' do
 
     describe 'basic' do
       it 'info' do
+        # see Rails::Rack::compute_tags
+        tags = log_tags.values.collect do |tag|
+          case tag
+          when Proc
+            tag.call(request)
+          when Symbol
+            request.send(tag)
+          else
+            tag
+          end
+        end
         logger[:abc] = 'xyz'
-        logger.tagged([request]) { logger.info('hello') }
+        logger.tagged(tags) { logger.info('hello') }
         expect(@my_logger.log).to eq([['foo', {
                                          abc: 'xyz',
                                          messages: ['hello'],
-                                         level: 'INFO',
+                                         severity: 'INFO',
                                          uuid: 'uuid_value',
                                          foo: 'foo_value'
                                        } ]])
         @my_logger.clear
-        logger.tagged([request]) { logger.info('world'); logger.info('bye') }
+        logger.tagged(tags) { logger.info('world'); logger.info('bye') }
         expect(@my_logger.log).to eq([['foo', {
                                          messages: ['world', 'bye'],
-                                         level: 'INFO',
+                                         severity: 'INFO',
                                          uuid: 'uuid_value',
                                          foo: 'foo_value'
                                        } ]])
@@ -73,8 +87,15 @@ EOF
     end
 
     describe 'frozen ascii-8bit string' do
-      it 'join messages' do
+      before do
         logger.instance_variable_set(:@messages_type, :string)
+      end
+
+      after do
+        logger.instance_variable_set(:@messages_type, :array)
+      end
+
+      it 'join messages' do
         ascii = "\xe8\x8a\xb1".force_encoding('ascii-8bit').freeze
         logger.tagged([request]) {
           logger.info(ascii)
@@ -85,10 +106,63 @@ EOF
       end
     end
 
+    describe 'Exception' do
+      it 'output message, class, backtrace' do
+        begin
+          3 / 0
+        rescue => e
+          logger.tagged([request]) {
+            logger.error(e)
+          }
+          expect(@my_logger.log[0][1][:messages][0]).
+            to match(%r|divided by 0 \(ZeroDivisionError\).*spec/logger_spec\.rb:|m)
+        end
+      end
+    end
+
+    describe 'Object' do
+      it 'output inspect' do
+        x = Object.new
+        logger.tagged([request]) {
+          logger.info(x)
+        }
+        expect(@my_logger.log[0][1][:messages][0]).to eq(x.inspect)
+      end
+    end
+
+    describe 'severity_key' do
+      describe 'not specified' do
+        it 'severity' do
+          logger.tagged([request]) { logger.info('hello') }
+          expect(@my_logger.log[0][1][:severity]).to eq('INFO')
+        end
+      end
+
+      describe 'severity_key: level' do
+        before do
+          @config_file = Tempfile.new('fluent-logger-config')
+          @config_file.close(false)
+          File.open(@config_file.path, 'w') {|f|
+            f.puts <<EOF
+test:
+  fluent_host: '127.0.0.1'
+  fluent_port: 24224
+  tag:         'foo'
+  severity_key: 'level'  # default severity
+EOF
+          }
+        end
+
+        it 'level' do
+          logger.tagged([request]) { logger.info('hello') }
+          expect(@my_logger.log[0][1][:level]).to eq('INFO')
+        end
+      end
+    end
   end
 
   describe "use ENV['FLUENTD_URL']" do
-    let(:fluentd_url) { "http://fluentd.example.com:42442/hoge?messages_type=string" }
+    let(:fluentd_url) { "http://fluentd.example.com:42442/hoge?messages_type=string&severity_key=level" }
 
     describe ".parse_url" do
       subject { described_class.parse_url(fluentd_url) }
@@ -96,6 +170,27 @@ EOF
       it { expect(subject['fluent_host']).to eq 'fluentd.example.com' }
       it { expect(subject['fluent_port']).to eq 42442 }
       it { expect(subject['messages_type']).to eq 'string' }
+      it { expect(subject['severity_key']).to eq 'level' }
+    end
+  end
+
+  describe 'flush_immediately' do
+    describe 'flush_immediately: true' do
+      it 'flushed' do
+        logger = ActFluentLoggerRails::Logger.new(config_file: File.new(@config_file.path),
+                                                  flush_immediately: true)
+        logger.info('Immediately!')
+        expect(@my_logger.log[0][1][:messages][0]).to eq('Immediately!')
+      end
+    end
+
+    describe 'flush_immediately: false' do
+      it 'flushed' do
+        logger = ActFluentLoggerRails::Logger.new(config_file: File.new(@config_file.path),
+                                                  flush_immediately: false)
+        logger.info('Immediately!')
+        expect(@my_logger.log).to eq(nil)
+      end
     end
   end
 end