+++ CHANGELOG.md
@@ -1,6 +1,30 @@
+## 0.3.1 / August 18 2016
+
+* Replace dependency from rails to railties and activesupport.
+
+## 0.3.0 / July 12 2016
+
+* Rails 5
+
+## 0.2.0 / Mar 20 2016
+
+* Add severity_key parameter. Is is The key of severity(DEBUG, INFO, WARN, ERROR).
+
+## 0.1.10 / Dec 23 2015
+
+* flush immediately.
+
+## 0.1.9 / Dec 16 2015
+
+ * Added settings: parameter to ActFluentLoggerRails::Logger.new.
+
+## 0.1.8 / Nov 14 2015
+
+ * Output Object#inspect if message is not String and not Exception.
+
 ## 0.1.7 / July 30 2015
 
-* Be able to log exceptions #15.
+ * Be able to log exceptions #15.
 
 ## 0.1.6 / March 20 2015